+++ packages/gatsby-recipes/src/validate-recipe.test.js
@@ -9,15 +9,23 @@ describe(`validate module validates recipes with resource declarations`, () => {
     ]
     const validationResponse = validateRecipe(recipe)
     expect(validationResponse[0].validationError).toBeTruthy()
-    //expect(validationResponse).toMatchSnapshot()
-    //expect(validationResponse[0].validationError).toMatchSnapshot()
+    expect(validationResponse).toMatchSnapshot()
+    expect(validationResponse[0].validationError).toMatchSnapshot()
   })
 
-  // it(`validates NPMPackage declarations`, () => {
-  //   const recipe = [{}, { NPMPackage: [{ namez: `wee-package` }] }]
-  //   const validationResponse = validateRecipe(recipe)
-  //   expect(validationResponse).toMatchSnapshot()
-  // })
+  it(`validates NPMPackage declarations`, () => {
+    const recipe = [{}, { NPMPackage: [{ namez: `wee-package` }] }]
+    const validationResponse = validateRecipe(recipe)
+    expect(validationResponse).toMatchSnapshot()
+  })
+
+  it(`returns errors for unknown resources`, () => {
+    const recipe = [{ Fake: [{}] }]
+
+    const result = validateRecipe(recipe)[0]
+
+    expect(result.validationError).toMatch(`Unknown resource Fake`)
+  })
 
   it(`returns empty array if there's no errors`, () => {
     const recipe = [