+++ packages/gatsby-transformer-sqip/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@2.3.1...gatsby-transformer-sqip@2.3.2) (2020-05-07)
+
+**Note:** Version bump only for package gatsby-transformer-sqip
+
+## [2.3.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@2.3.0...gatsby-transformer-sqip@2.3.1) (2020-05-05)
+
+**Note:** Version bump only for package gatsby-transformer-sqip
+
+# [2.3.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@2.2.7...gatsby-transformer-sqip@2.3.0) (2020-04-27)
+
+**Note:** Version bump only for package gatsby-transformer-sqip
+
+## [2.2.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@2.2.6...gatsby-transformer-sqip@2.2.7) (2020-04-24)
+
+**Note:** Version bump only for package gatsby-transformer-sqip
+
 ## [2.2.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@2.2.5...gatsby-transformer-sqip@2.2.6) (2020-04-17)
 
 ### Bug Fixes